+++ src/main/java/com/pullanner/api/controller/article/dto/ArticleResponses.java
@@ -1,4 +1,4 @@
-package com.pullanner.domain.article.dto;
+package com.pullanner.api.controller.article.dto;
 
 import com.pullanner.domain.article.entity.Article;
 import java.util.List;
