+++ src/main/java/com/pullanner/api/controller/article/dto/ArticleUpdateRequest.java
@@ -1,4 +1,4 @@
-package com.pullanner.domain.article.dto;
+package com.pullanner.api.controller.article.dto;
 
 import lombok.Getter;
 import lombok.NoArgsConstructor;
